fix: apply code-review findings across core, npm layer, and Electron example
Correctness (Rust core):
- wasm segment() now uses the constructor's config instead of silently
  falling back to defaults — push() and segment() on one instance agreed
  to disagree before.
- Pinned negativeThreshold no longer overrides the entry threshold. The
  Swift original derives entry = negative + offset because its entry
  threshold lives on a different struct; porting that into a single config
  silently discarded a user-set threshold. Documented deviation; entry is
  always `threshold`, pinning only moves the exit.
- Config validation (NaN, negatives, thresholds out of [0,1], inverted
  hysteresis) — a NaN threshold previously disabled the state machine
  silently, trivially reachable from JS.
- Streaming counters u64: usize is 32-bit on wasm32 and a continuous mic
  session wraps after ~74 h, corrupting temp_end arithmetic. Event
  sampleIndex surfaces as f64 in JS (exact, no BigInt).
- VadStreamer::push drains the consumed prefix even on error so a retry
  never replays frames the LSTM state already saw.
- probabilities() zero-pads and scores the trailing partial frame (as
  FluidAudio does) instead of dropping up to 511 samples of tail speech.
- streaming_state_machine takes &mut state — kills a per-frame deep clone
  and a stale 'for unit testing' doc.

npm layer:
- load() no longer caches a rejected promise forever (one transient wasm
  fetch failure used to brick every future createVad()).
- VadEvent/Segment exported at runtime (index.d.ts already claimed them;
  importing threw SyntaxError).
- MicVad: start/stop race fixes (double-start double-acquired the mic;
  stop during in-flight start left capture running), late worklet messages
  silenced after stop, utterance audio now trimmed to exactly
  [startTime, endTime] via absolute sample indexing, workletUrl option for
  bundlers that don't rewrite new URL(), frame size flows from the wasm
  constant through processorOptions (single source of truth).
- package.json sideEffects lists worklet.js (was 'false' — bundlers could
  tree-shake the registerProcessor side effect away).
- Type docs: which options apply to segment() vs streaming; validation.

Electron example / CI / scripts:
- main.js: all rejections funnel to app.exit(1) (an unhandled rejection
  used to hang CI to the 6h job timeout); window-all-closed can no longer
  convert an incomplete smoke into exit 0; renderer watchdog starts after
  page load; smoke uses 64 whole frames (was a fractional 62.5).
- ci.yml: timeout-minutes on all jobs.
- prepare_model.py: atomic download (temp + rename).
- scripts/make_testdata.py: the previously-referenced-but-missing fixture
  generator; ORT reference regenerated to include the padded tail frame.

+++ tests/streaming.rs
@@ -31,7 +31,10 @@ fn streamer_buffers_arbitrary_chunk_sizes() {
         results.extend(streamer.push(chunk).expect("push"));
     }
     assert_eq!(results.len(), samples.len() / FRAME_SIZE);
-    assert_eq!(streamer.processed_samples(), results.len() * FRAME_SIZE);
+    assert_eq!(
+        streamer.processed_samples(),
+        (results.len() * FRAME_SIZE) as u64
+    );
 }
 
 #[test]
